Fixe #2711: Removed unnecessary boxing on Float value

+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/utils/QuantityParserUtil.java
@@ -115,7 +115,7 @@ public class QuantityParserUtil {
     public static Float getFloatValue(String initText, EntryFormat floatFormat) {
         Double result = getDoubleValue(initText, floatFormat);
         if (result != null) {
-            return Float.valueOf(result.floatValue());
+            return result.floatValue();
         }
         return null;
     }
